requirements.txt
fastai
pillow
streamlit
